fix : 돈길계약하기 이후 모달에는 사인 이미지 자체로 사용 (s3url ㄴㄴ)
