+++ src/components/home/create/steps/Step3.tsx
@@ -13,7 +13,6 @@ import {
 import SheetButton from '@components/common/buttons/SheetButton';
 import InputForm from '@components/common/forms/InputForm';
 import useBottomSheetOutSideRef from '@lib/hooks/useBottomSheetOutSideRef';
-import getCommaThreeDigits from '@lib/utils/get/getCommaThreeDigits';
 import ContractSheet from '@components/common/bottomSheets/contractSheet/ContractSheet';
 import { useQueryClient } from 'react-query';
 import queryKeys from '@lib/constants/queryKeys';
