+++ src/components/walk/SwipeToWalk.tsx
@@ -7,7 +7,6 @@ import { ReactComponent as Stamp } from '@assets/illusts/walk/achieved-stamp.svg
 import Slider from 'rc-slider';
 import 'rc-slider/assets/index.css';
 import { TInterestRate } from '@lib/types/IInterestRate';
-import getCommaThreeDigits from '@lib/utils/get/getCommaThreeDigits';
 import { useMutation, useQueryClient } from 'react-query';
 import challengeAPI from '@lib/apis/challenge/challengeAPI';
 import queryKeys from '@lib/constants/queryKeys';
