--- src/lib/utils/get/getCommaThreeDigits.ts
@@ -1,5 +1,0 @@
-const getCommaThreeDigits = (number: number) => {
-  return number.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-};
-
-export default getCommaThreeDigits;
